Chdir instead of using `-p' when the path of the app is a directory
This change will fix an issue when deploying multiple applications using
one manifest file. Currently the `cf' cli will complain with the
followign error message:

Command line flags (except -f) cannot be applied when pushing multiple
apps from a manifest file

This commit modifies the logic to chdir to the app path if it is a
directory, otherwise it reverts to the old behavior and passes `-p' flag
to the cli.