Fixes #29 allow setting environment variables including CF_DIAL_TIMEOUT
**Why**

Using the `cf` resource in a pipeline which has a high latency
connection to the CloudFoundry instance fails and the solution is to set
the `CF_DIAL_TIMEOUT` to a higher value (15s is recommended). For
instance, this occurs running a Concourse pipeline in AWS Sydney and
targeting PWS in the US.
By extension, there is no mechanism for setting any of the `CF_`
environment variables (or other environment variables if they were
needed).

**How**

The design of the solution has taken a general 'allow shell environment
setting' approach as opposed to a 'create individual config params for
all CF_ switches'. This is to remove the burden of keeping this resource
in sync with the `cf` command line tool while giving the benefit of easy
environment config.

The solution adds a `CommandEnvironmentVariables` struct to the
`Source` struct, allowing a map of environment key-value pairs to be
added in to the Source Configuration YAML.

A new CfEnvironment struct is used to hold the internal model of the
environment variables. There are constructors to pre-populated with
environment variables or not. Both constructors pre-populate
`CF_COLOR=true` as before. The CloudFoundry struct is now composed of
a CfEnvironment.

The CfEnvironment uses a `map[string]string' but the JSON Unmarshalling
uses a `map[string]interface{}` for ease of type safety, so there is a
bit of `fmt.Sprintf("%v", v)` code for handling the conversion.

The `CommandEnvironmentVariables` is populated by the `main` during
JSON wrangling and then added to the `cloudFoundry`.
The environment is finally set up (as before) in the `CloudFoundry.cf`
method.

**Effects**

Any environment variables may be now set for the runtime environment of
the `cf` command via the Source Configuration YAML. Existing set
variable can be overridden. This is why a `map` was used internally
rather than `[]string` of `KEY=VAL` that os.Environ() returns and
`cmd.Env` takes.